Refactor CustomMainMenu -> mainmenu.PSSMainMenu
