+++ src/main/kotlin/me/partlysanestudios/partlysaneskies/PartlySaneSkies.kt
@@ -64,12 +64,13 @@ import me.partlysanestudios.partlysaneskies.features.farming.garden.SkymartValue
 import me.partlysanestudios.partlysaneskies.features.farming.garden.VisitorLogbookStats
 import me.partlysanestudios.partlysaneskies.features.farming.garden.VisitorTradeValue
 import me.partlysanestudios.partlysaneskies.features.foraging.TreecapitatorCooldown
-import me.partlysanestudios.partlysaneskies.features.gui.CustomMainMenu
+import me.partlysanestudios.partlysaneskies.features.gui.PSSMainMenu
 import me.partlysanestudios.partlysaneskies.features.gui.RefreshKeybinds
 import me.partlysanestudios.partlysaneskies.features.gui.hud.CooldownHud
 import me.partlysanestudios.partlysaneskies.features.gui.hud.LocationBannerDisplay
 import me.partlysanestudios.partlysaneskies.features.gui.hud.rngdropbanner.DropBannerDisplay
 import me.partlysanestudios.partlysaneskies.features.gui.hud.rngdropbanner.DropWebhook
+import me.partlysanestudios.partlysaneskies.features.gui.mainmenu.PSSMainMenu
 import me.partlysanestudios.partlysaneskies.features.information.WikiArticleOpener
 import me.partlysanestudios.partlysaneskies.features.mining.MiningEvents
 import me.partlysanestudios.partlysaneskies.features.mining.PickaxeWarning
@@ -165,7 +166,7 @@ class PartlySaneSkies {
             PublicDataManager.getFile("main_menu.json")
         }.start()
         Thread {
-            CustomMainMenu.loadFunFact()
+            PSSMainMenu.loadFunFact()
         }.start()
 
         // Loads extra json data
@@ -248,7 +249,7 @@ class PartlySaneSkies {
         registerEvent(BitsShopValue)
         registerEvent(SkymartValue)
         registerEvent(VisitorTradeValue)
-        registerEvent(CustomMainMenu.Companion)
+        registerEvent(PSSMainMenu)
         registerEvent(WrongToolCropWarning.CropToolData)
         registerEvent(PetAlert)
